+++ src/gui/auxiliary/typed_formula.hpp
@@ -245,10 +245,10 @@ inline color_t typed_formula<color_t>::execute(wfl::variant& v) const
 	const int alpha = result.size() == 4 ? result[3].as_int() : ALPHA_OPAQUE;
 
 	return color_t(
-		result.at(0).as_int(),
-		result.at(1).as_int(),
-		result.at(2).as_int(),
-		alpha
+		static_cast<uint8_t>(result.at(0).as_int()),
+		static_cast<uint8_t>(result.at(1).as_int()),
+		static_cast<uint8_t>(result.at(2).as_int()),
+		static_cast<uint8_t>(alpha)
 	);
 }
